UnitTestJUnitMockito-lab11/src/main/java/edu/miu/cs/cs489appsd/ArrayFlattener.java
package edu.miu.cs.cs489appsd;

import java.util.ArrayList;
import java.util.List;

public class ArrayFlattener {
    public int[] flattenArray(int[][] nestedArray) {
        if (nestedArray == null) {
            return new int[0];
        }

        List<Integer> flattenedList = new ArrayList<>();
        for (int[] innerArray : nestedArray) {
            for (int num : innerArray) {
                flattenedList.add(num);
            }
        }

        int[] flattenedArray = new int[flattenedList.size()];
        for (int i = 0; i < flattenedList.size(); i++) {
            flattenedArray[i] = flattenedList.get(i);
        }

        return flattenedArray;
    }
}
